[fix] up some outdated code that made somewhat no sense, file structure in terms of how code is organized, and the naming choice for what was previously a form
